Refactor: Moves quest's InOutOctreeValidator into its own file
